removed supra class SynthPop, renamed EpiSynthPop to SynthPop
